+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+
+`0.5`_ (2016-07-07)
+~~~~~~~~~~~~~~~~~~~
+
 - Fixed a crash where apps without ``required_fields`` could not be
   saved.
 - Added a template snippet based renderer for plugins.
@@ -78,4 +82,5 @@ Changed / Deprecated / Removed / Fixed / Security
 .. _0.2: https://github.com/matthiask/feincms3/compare/0.1...0.2
 .. _0.3: https://github.com/matthiask/feincms3/compare/0.2...0.3
 .. _0.4: https://github.com/matthiask/feincms3/compare/0.3...0.4
-.. _Next version: https://github.com/matthiask/feincms3/compare/0.4...master
+.. _0.5: https://github.com/matthiask/feincms3/compare/0.4...0.5
+.. _Next version: https://github.com/matthiask/feincms3/compare/0.5...master